test(auto-reply): split model-selection directive tests under the file-size limit
directive-handling.mixed-inline.test.ts grew past the 1000-line lint limit;
move the model-scope scope-resolution table and the authoritative
model-lock rejection tests into a dedicated file with the same mock setup.
